♻️ refactor(aow-hlo-create-result) P2-1876: Refactor and modularize report result modal
Extracted the report result modal into a new AowHloCreateModalComponent for better modularity and maintainability. Updated the table component to use the new modal, moved related SCSS, and enhanced the modal to support dynamic entity and project selection. Added new signals and methods in EntityAowService to manage selected entities and W3/bilateral projects, and extended ResultsApiService with a new API call for bilateral projects.